+++ hypothesis_sqlalchemy/scheme.py
@@ -1,0 +1,9 @@
+from .core import (column as _column,
+                   column_type as _column_type,
+                   columns as _columns,
+                   table as _table)
+
+column_types = _column_type.instances
+columns = _column.instances
+columns_lists = _columns.lists
+tables = _table.instances
