--- hypothesis_sqlalchemy/tabular/__init__.py
@@ -1,2 +1,0 @@
-from . import records
-from .base import factory
